refactor: OrderValidator 추상화하여 OrderTable 의존성 제거
